seperated data and fuctionality
